requirements.txt
fastapi
uvicorn
pyserial
requests